refactor(profile-page): merge visitor and member components
